refactor: Organizar componente
