+++ frontend/src/components/camposDeFormulario/TextEditor/index.ts
@@ -1,0 +1,3 @@
+import TextEditor from './TextEditor.vue';
+
+export default TextEditor;
